awesome: Fix multimedia keys and add pause/nxt/prv
A handler for multimedia keys would be added FOR EACH TAG, which would
result in *way* too many events being triggered.

Added play/pause/next/previous support through `playerctrl`.

Fixed display bug where there would be a small area at the bottom of the
screen and on the right that was not covered by the terminal. This was
by design to fit the amount of columns/rows, but I prefer the empty
space added by not honoring the size hints.